Add call to flash_attn varlen kernel

+++ vllm/worker/model_runner.py
@@ -257,7 +257,7 @@ class ModelRunner:
         # Cumulative index of each prompt. [prompt_lens + 1]
         # [0, 0+1th, 0+1th+2nd, ...]
         start_loc_tensor = torch.zeros(prompt_lens_tensor.shape[0] + 1,
-                                       dtype=torch.long,
+                                       dtype=torch.int32,
                                        device=self.device)
 
         torch.cumsum(prompt_lens_tensor,
